♻️ set passphrase in layout
